requirements.txt
django==5.2.1
djangorestframework==3.16.0
pandas==2.2.3
openpyxl==3.1.3
xlrd==2.0.1
pyxform==3.0.1
gunicorn==21.2.0
whitenoise==6.5.0
psycopg2-binary==2.9.7
elementpath==5.0.1
ruff==0.11.11
